requirements.txt
Flask==2.0.2
requests==2.26.0
ptvsd==4.3.2 # Required for debugging.
Pillow==8.4.0
